requirements.txt
langchain>=0.3.10
langchain-openai>=0.2.6
pygdbmi>=0.11.0.0
prompt_toolkit>=3.0.47
rich>=13.7.1
python-dotenv>=1.0.1
pydantic>=2.7.4
